Add transaction details
