+++ Sources/PulseUI/Views/KeyValueSectionView+Extensions.swift
@@ -111,10 +111,18 @@ extension KeyValueSectionViewModel {
         )
     }
 
-#warning("TODO: remove?")
-
 #if os(iOS) || os(macOS) || os(tvOS)
-    static func makeTiming(for transaction: NetworkTransactionMetricsEntity) -> KeyValueSectionViewModel {
+    static func makeDetails(for transaction: NetworkTransactionMetricsEntity) -> [KeyValueSectionViewModel] {
+        return [
+            makeTiming(for: transaction),
+            makeTransferSection(for: transaction),
+            makeProtocolSection(for: transaction),
+            makeMiscSection(for: transaction),
+            makeSecuritySection(for: transaction)
+        ].compactMap { $0 }
+    }
+
+    private static func makeTiming(for transaction: NetworkTransactionMetricsEntity) -> KeyValueSectionViewModel {
         let timeFormatter = DateFormatter()
         timeFormatter.locale = Locale(identifier: "en_US")
         timeFormatter.dateFormat = "hh:mm:ss.SSS"
@@ -152,7 +160,55 @@ extension KeyValueSectionViewModel {
         }
         return KeyValueSectionViewModel(title: "Timing", color: .orange, items: items)
     }
+
+    private static func makeTransferSection(for metrics: NetworkTransactionMetricsEntity) -> KeyValueSectionViewModel? {
+        let transferSize = metrics.transferSize
+        return KeyValueSectionViewModel(title: "Data Transfer", color: .primary, items: [
+            ("Request Headers", formatBytes(transferSize.requestHeaderBytesSent)),
+            ("Request Body", formatBytes(transferSize.requestBodyBytesBeforeEncoding)),
+            ("Request Body (Encoded)", formatBytes(transferSize.requestBodyBytesSent)),
+            ("Response Headers", formatBytes(transferSize.responseHeaderBytesReceived)),
+            ("Response Body", formatBytes(transferSize.responseBodyBytesReceived)),
+            ("Response Body (Decoded)", formatBytes(transferSize.responseBodyBytesAfterDecoding))
+        ])
+    }
+
+    private static func makeProtocolSection(for metrics: NetworkTransactionMetricsEntity) -> KeyValueSectionViewModel? {
+        KeyValueSectionViewModel(title: "Protocol", color: .primary, items: [
+            ("Network Protocol", metrics.networkProtocol),
+            ("Remote Address", metrics.remoteAddress),
+            ("Remote Port", metrics.remotePort > 0 ? String(metrics.remotePort) : nil),
+            ("Local Address", metrics.localAddress),
+            ("Local Port", metrics.localPort > 0 ? String(metrics.localPort) : nil)
+        ])
+    }
+
+    private static func makeSecuritySection(for metrics: NetworkTransactionMetricsEntity) -> KeyValueSectionViewModel? {
+        guard let suite = metrics.negotiatedTLSCipherSuite,
+              let version = metrics.negotiatedTLSProtocolVersion else {
+            return nil
+        }
+        return KeyValueSectionViewModel(title: "Security", color: .primary, items: [
+            ("Cipher Suite", suite.description),
+            ("Protocol Version", version.description)
+        ])
+    }
+
+    private static func makeMiscSection(for metrics: NetworkTransactionMetricsEntity) -> KeyValueSectionViewModel? {
+        KeyValueSectionViewModel(title: "Characteristics", color: .primary, items: [
+            ("Cellular", metrics.isCellular.description),
+            ("Expensive", metrics.isExpensive.description),
+            ("Constrained", metrics.isConstrained.description),
+            ("Proxy Connection", metrics.isProxyConnection.description),
+            ("Reused Connection", metrics.isReusedConnection.description),
+            ("Multipath", metrics.isMultipath.description)
+        ])
+    }
 #endif
+}
+
+private func formatBytes(_ count: Int64) -> String {
+    ByteCountFormatter.string(fromByteCount: count)
 }
 
 #warning("TODO: where else should we use this? definitely for all exports")
